Add support transactional emails.
Give the user of this library the ability to distinguish transactional
emails from normal emails that would fall under the CAN-SPAM Act. This
then allows the email provider to treat these emails differently as
well.

See the FTC's CAN-SPAM Act compliance guide for further information on
what a transactional email is, how it differs from normal emails a
business sends, and the rules are that they both must follow at
https://www.ftc.gov/tips-advice/business-center/guidance/can-spam-act-compliance-guide-business