Use `ZC_EVENTS_SETTINGS` first and django second.
If using django settings only, and the settings file imports a model,
django will fail hard on startup.

This could be common scenario for the JOB_MAPPING setting which takes a
key of a string and the value is the function to execute. Thus, if the
file that the function lives in imports a django model, the app will not
be initialized yet.