Fix bug in determining job mappings.
Fixes #52

Fix the bug in determining job mappings by doing a collection check
instead of just an `hasattr` check. In addition, provide an example of
how to work with `query_params` in an APIView.

Finally, add the python 2 django worker to the docker-compose file, and
bump the version for a new release.